comfyui-runninghub/pyproject.toml
[tool.poetry]
name = "comfyui-runninghub"
version = "0.1.0"
description = "FastAPI microservices to interact with Runninghub AI image generation"
authors = ["JOTO Team <dev@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
httpx = "^0.27.2"
python-multipart = "^0.0.9"
tenacity = "^9.0.0"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


